requirements.txt
matplotlib==3.8.1
mediapipe==0.10.5
numpy==1.25.2
odfpy==1.4.1
opencv_contrib_python==4.8.1.78
opencv_python==4.8.0.76
pandas==2.1.4
PySide6==6.5.2
PySide6_Addons==6.5.2
PySide6_Essentials==6.5.2
